+++ packages/template-retail-react-app/app/components/_app-config/index.test.js
@@ -52,7 +52,6 @@ describe('AppConfig', () => {
 
     afterAll(() => {
         window.localStorage.setItem.mockRestore()
-        global.fetch.mockClear()
         global.fetch = originalFetch
     })